Native 1.2: move common enum types to native.* package, do TODOs
