+++ native/native.go
@@ -1,0 +1,5 @@
+// Package native provides OpenRTB Dynamic Native Ads API Specification Version 1.2
+// section 7 Reference Lists/Enumerations types/constants:
+// https://www.iab.com/guidelines/real-time-bidding-rtb-project/
+// https://iabtechlab.com/wp-content/uploads/2016/07/OpenRTB-Native-Ads-Specification-Final-1.2.pdf
+package native
